Refactor detector and source_post_processors.
